+++ src/intelli/gui/graphics/commentobject.h
@@ -126,11 +126,11 @@ private:
     /// anchor object and cannot be moved unless its uncollapsed.
     QPointer<GraphicsObject const> m_anchor;
     /// Main widget
-    QGraphicsProxyWidget* m_proxyWidget;
+    QGraphicsProxyWidget* m_proxyWidget = {};
     /// Overlay widget to suppress mouse and key event to the main widget
-    Overlay* m_overlay;
+    Overlay* m_overlay = {};
     /// Comment editor
-    QTextEdit* m_editor;
+    QTextEdit* m_editor = {};
 
     QRectF resizeHandleRect() const;
 
@@ -141,9 +141,6 @@ private:
     void selectColor();
 
     void onColorChanged();
-
-    QString m_color;
-    QString m_bgcolor;
 
 private slots: