+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Selected nodes can now be "cut" using the corresponding shortcut (usually `Ctrl+X`). Cut-nodes are "greyed-out", similar to how cutting works for files and folders. Once the selection is pasted, the cut-objects are deleted. - #283
-- Comments background color and text color can now be changed using a dialog window accessible from the context menu
+- Comments background color and text color can now be changed using its context menu. - #326
+- The border of comments may be hidden using its context menu. - #326
+- The text of comments may now be centered using its context menu. - #326
 
 ### Changed